objecttypeCB type float to int change

+++ JNSEngine/Engine_Source/jnsRenderer.h
@@ -33,7 +33,7 @@ namespace renderer
 
 	CBUFFER(ObjectTypeMoveCB, CBSLOT_MOVE)
 	{	
-		float mtype;
+		int mtype;
 		Vector3 mTime;
 	};
